Move operator utils from schema_metadata.js to metabase-lib
