+++ frontend/src/metabase-lib/lib/metadata/Table.ts
@@ -4,9 +4,9 @@
 // NOTE: this needs to be imported first due to some cyclical dependency nonsense
 import Question from "../Question"; // eslint-disable-line import/order
 import { singularize } from "metabase/lib/formatting";
-import { getAggregationOperators } from "metabase/lib/schema_metadata";
 import type { TableId } from "metabase-types/types/Table";
 import { isVirtualCardId } from "metabase/lib/saved-questions/saved-questions";
+import { getAggregationOperators } from "metabase-lib/lib/operators/utils";
 import { createLookupByProperty, memoizeClass } from "metabase-lib/lib/utils";
 import Base from "./Base";
 import type Metadata from "./Metadata";
