+++ frontend/src/metabase-lib/lib/queries/StructuredQuery.ts
@@ -10,7 +10,6 @@ import {
   format as formatExpression,
   DISPLAY_QUOTES,
 } from "metabase/lib/expressions/format";
-import { isCompatibleAggregationOperatorForField } from "metabase/lib/schema_metadata";
 import {
   StructuredQuery as StructuredQueryObject,
   Aggregation,
@@ -32,6 +31,7 @@ import {
   isVirtualCardId,
   getQuestionIdFromVirtualTableId,
 } from "metabase/lib/saved-questions";
+import { isCompatibleAggregationOperatorForField } from "metabase-lib/lib/operators/utils";
 import { TYPE } from "metabase-lib/lib/types/constants";
 import { fieldRefForColumn } from "metabase-lib/lib/queries/utils/dataset";
 import { isSegment } from "metabase-lib/lib/queries/utils/filter";
