+++ frontend/src/metabase-lib/lib/queries/drills/summarize-column-drill.js
@@ -2,7 +2,7 @@ import _ from "underscore";
 import {
   getAggregationOperator,
   isCompatibleAggregationOperatorForField,
-} from "metabase/lib/schema_metadata";
+} from "metabase-lib/lib/operators/utils";
 import { isa } from "metabase-lib/lib/types/utils/isa";
 import { TYPE } from "metabase-lib/lib/types/constants";
 import { fieldRefForColumn } from "metabase-lib/lib/queries/utils/dataset";
