+++ frontend/src/metabase-lib/lib/queries/structured/Filter.ts
@@ -9,7 +9,7 @@ import {
 } from "metabase-types/types/Query";
 import { FilterOperator } from "metabase-types/types/Metadata";
 import { isExpression } from "metabase/lib/expressions";
-import { getFilterArgumentFormatOptions } from "metabase/lib/schema_metadata";
+import { getFilterArgumentFormatOptions } from "metabase-lib/lib/operators/utils";
 import {
   generateTimeFilterValuesDescriptions,
   getRelativeDatetimeField,
