+++ frontend/src/metabase/parameters/components/widgets/ParameterFieldWidget/ParameterFieldWidget.jsx
@@ -7,15 +7,15 @@ import _ from "underscore";
 import FieldValuesWidget from "metabase/components/FieldValuesWidget";
 import { deriveFieldOperatorFromParameter } from "metabase/parameters/utils/operators";
 import {
-  getFilterArgumentFormatOptions,
-  isEqualsOperator,
-  isFuzzyOperator,
-} from "metabase/lib/schema_metadata";
-import {
   WidgetRoot,
   Footer,
   UpdateButton,
 } from "metabase/parameters/components/widgets/Widget.styled";
+import {
+  getFilterArgumentFormatOptions,
+  isEqualsOperator,
+  isFuzzyOperator,
+} from "metabase-lib/lib/operators/utils";
 
 import { normalizeValue } from "./normalizeValue";
