+++ frontend/src/metabase/parameters/utils/operators.ts
@@ -3,7 +3,7 @@ import { UiParameter } from "metabase/parameters/types";
 import {
   doesOperatorExist,
   getOperatorByTypeAndName,
-} from "metabase/lib/schema_metadata";
+} from "metabase-lib/lib/operators/utils";
 import { NUMBER, STRING, PRIMARY_KEY } from "metabase-lib/lib/types/constants";
 import { PARAMETER_OPERATOR_TYPES } from "../constants";
 import { getParameterType, getParameterSubType } from "./parameter-type";
