+++ frontend/src/metabase/query_builder/components/Filter.jsx
@@ -2,13 +2,13 @@
 import React from "react";
 
 import { t, ngettext, msgid } from "ttag";
-import { getFilterArgumentFormatOptions } from "metabase/lib/schema_metadata";
 
 import Value from "metabase/components/Value";
 
 import { color } from "metabase/lib/colors";
 
 import ViewPill from "metabase/query_builder/components/view/ViewPill";
+import { getFilterArgumentFormatOptions } from "metabase-lib/lib/operators/utils";
 import { generateTimeFilterValuesDescriptions } from "metabase-lib/lib/queries/utils/query-time";
 import { hasFilterOptions } from "metabase-lib/lib/queries/utils/filter";
 import Dimension from "metabase-lib/lib/Dimension";
